:recycle: refactor(BTS-300): 중복 제거 및 사이드바 아이콘 분리
