+++ src/shared/components/icons/index.ts
@@ -1,0 +1,5 @@
+export { PlusIcon } from './plus-icon';
+export { SettingIcon } from './setting-icon';
+export { TextIcon } from './text-icon';
+export { ListIcon } from './list-icon';
+export { HomeIcon } from './home-icon';
