+++ src/shared/components/icons/plus-icon.tsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+export const PlusIcon = () => {
+  return (
+    <svg
+      width="16"
+      height="16"
+      viewBox="0 0 16 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M7.5 13.5L7.5 8L13 8M7.5 2.5L7.49948 8.00052L2 8"
+        stroke="#7C7C7C"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+};
